рефакторит функции даты и времени
