+++ src/mock/point.js
@@ -1,43 +1,6 @@
-// eslint-disable-next-line no-unused-vars
-import dayjs from 'dayjs';
-
+import {generateDate, getRandomIntInclusive, getRandomInteger} from '../utils.js';
 
 const MAXIMUM_NUMBER_OF_SENTENCES = 5;
-
-
-/**
- * Функция из интернета по генерации случайного числа из диапазона
- * Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
- * @param a - минимальное значение
- * @param b - максимальное значение
- * @returns {number}
- */
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-
-/**
- * Функция, возвращающая случайное целое число из переданного диапазона включительно.
- * Результат: целое число из диапазона "от...до".
- * Источник: https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
- *
- * @param min - минимальное значекние диапозона
- * @param max - максимальное значение диапозона
- */
-const getRandomIntInclusive = (min, max) => {
-  const minNumber = Math.ceil(min);
-  const maxNumber = Math.floor(max);
-  if (minNumber >= maxNumber) {
-    return 'Введены неверные значения';
-  }
-
-  // Максимум и минимум включаются
-  return Math.floor(Math.random() * (maxNumber - minNumber + 1)) + minNumber;
-};
 
 
 /**
@@ -81,18 +44,6 @@ const generatePointType = () => {
   const randomIndex = getRandomInteger(0, pointType.length - 1);
 
   return pointType[randomIndex];
-};
-
-
-/**
- * Генерирует случайную дату в заданном диапазоне
- * @returns {Date}
- */
-const generateDate = () => {
-  const maxMinuteGap = 4320;
-  const daysGap = getRandomInteger(0, maxMinuteGap);
-
-  return dayjs().add(daysGap, 'minute').toDate();
 };
